REFS: #243 and fixes #243 (Changes in Myntra Project React)

+++ Modules/Project/Nihar Sodhaparmar/React/myntra/src/pages/BagPage.js
@@ -1,11 +1,12 @@
 import React, { useEffect, useState } from "react";
 import BagService from "../services/BagService";
 import OrderService from "../services/OrderService";
-import { getToken, getUserId, removeUserSession } from "../Utils/Storage";
+import { getToken, getUserId, removeUserSession } from "../utils/Storage";
 import EmptyBanner from "../components/EmptyBanner";
 import BagItem from "../components/BagPage/BagItem";
 import Loading from "../components/Loading";
 import Navbar from "../components/Navbar";
+import { NotificationManager } from "react-notifications";
 
 export default function BagPage(props) {
   const [bagItems, setBagItems] = useState([]);
@@ -46,17 +47,20 @@ export default function BagPage(props) {
   }, []);
 
   const removeFromBag = async (bagItemId) => {
-    try {
-      setLoading(true);
-      await BagService.deleteBagItem(bagItemId, getToken());
-      setLoading(false);
-      getData();
-    } catch (error) {
-      if (error.response.status === 401) {
-        props.history.push("/login");
-        removeUserSession();
+    if (window.confirm("Are you sure you want to remove item from bag?")) {
+      try {
+        setLoading(true);
+        await BagService.deleteBagItem(bagItemId, getToken());
+        setLoading(false);
+        NotificationManager.error("Item removed from bag", "", 2000);
+        getData();
+      } catch (error) {
+        if (error.response.status === 401) {
+          props.history.push("/login");
+          removeUserSession();
+        }
+        console.error(error);
       }
-      console.error(error);
     }
   };
 
@@ -87,32 +91,35 @@ export default function BagPage(props) {
   };
 
   const placeOrder = async () => {
-    try {
-      setLoading(true);
-      let tempProducts = bagItems.map((product) => {
-        let newProduct = {
-          product: product._id,
-          size: product.size,
-          quantity: product.quantity,
-        };
+    if (window.confirm("Are you want to place an order?")) {
+      try {
+        setLoading(true);
+        let tempProducts = bagItems.map((product) => {
+          let newProduct = {
+            product: product._id,
+            size: product.size,
+            quantity: product.quantity,
+          };
 
-        return newProduct;
-      });
-      let data = {
-        customer: getUserId(),
-        products: tempProducts,
-      };
-      await OrderService.addOrder(data, getToken());
-      setLoading(false);
-      getData();
-      props.history.push("/orders");
-    } catch (error) {
-      setLoading(false);
-      if (error.response.status === 401) {
-        props.history.push("/login");
-        removeUserSession();
+          return newProduct;
+        });
+        let data = {
+          customer: getUserId(),
+          products: tempProducts,
+        };
+        await OrderService.addOrder(data, getToken());
+        setLoading(false);
+        getData();
+        NotificationManager.success("Order placed successfully", "", 2000);
+        props.history.push("/orders");
+      } catch (error) {
+        setLoading(false);
+        if (error.response.status === 401) {
+          props.history.push("/login");
+          removeUserSession();
+        }
+        console.error(error);
       }
-      console.error(error);
     }
   };
 
@@ -121,7 +128,17 @@ export default function BagPage(props) {
 
   if (bagItems.length >= 1) {
     bagItems.map((product) => {
-      total += parseInt(product.quantity) * parseFloat(product.price);
+      if (product.offer > 0) {
+        total +=
+          parseInt(product.quantity) *
+          (parseInt(product.price) -
+            parseInt(
+              (parseInt(product.price) * parseInt(product.offer)) / 100
+            ));
+      } else {
+        total += parseInt(product.quantity) * parseFloat(product.price);
+      }
+
       totalItems += parseInt(product.quantity);
       return null;
     });
@@ -152,7 +169,7 @@ export default function BagPage(props) {
         <div className="text-center pt-4 category-header text-capitalize">
           bag items
         </div>
-        <div className="container my-5" style={{ width: "fit-content" }}>
+        <div className="container my-4" style={{ width: "fit-content" }}>
           <ul className="list-group">
             {bagItems.map((item, index) => {
               return (
